restoring tables
