+++ src/Azure/Storage.Net.Microsoft.Azure.Storage.Tables/Tables.csproj
@@ -2,25 +2,22 @@
 
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
-    <AssemblyName>Storage.Net.Microsoft.Azure.Storage.Files</AssemblyName>
-    <RootNamespace>Storage.Net.Microsoft.Azure.Storage.Files</RootNamespace>
-     <LangVersion>latest</LangVersion>
      <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-     <WarningsAsErrors />
      <GenerateDocumentationFile>true</GenerateDocumentationFile>
+     <WarningsAsErrors />
+     <AssemblyName>Storage.Net.Microsoft.Azure.Storage.Tables</AssemblyName>
+     <RootNamespace>Storage.Net.Microsoft.Azure.Storage.Tables</RootNamespace>
+
   </PropertyGroup>
 
    <PropertyGroup Condition="'$(Configuration)' == 'Release'">
       <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
    </PropertyGroup>
 
-
-   <ItemGroup>
-      <PackageReference Include="Microsoft.Azure.Storage.File" Version="11.1.0" />
-   </ItemGroup>
-
    <ItemGroup>
      <ProjectReference Include="..\..\Storage.Net\Storage.Net.csproj" />
+      <PackageReference Include="Microsoft.Azure.Cosmos.Table" Version="1.0.5" />
    </ItemGroup>
+
 
 </Project>
