+++ src/Azure/Storage.Net.Microsoft.Azure.Storage/AzureStorageModule.cs
@@ -2,7 +2,6 @@
 using Storage.Net.ConnectionString;
 using Storage.Net.KeyValue;
 using Storage.Net.Messaging;
-using Storage.Net.Microsoft.Azure.Storage.KeyValue;
 using Storage.Net.Microsoft.Azure.Storage.Messaging;
 
 namespace Storage.Net.Microsoft.Azure.Storage
@@ -16,21 +15,6 @@ namespace Storage.Net.Microsoft.Azure.Storage
 
       public IKeyValueStorage CreateKeyValueStorage(StorageConnectionString connectionString)
       {
-         if(connectionString.Prefix == Constants.AzureTablesConnectionPrefix)
-         {
-            if(bool.TryParse(connectionString.Get(Constants.UseDevelopmentStorage), out bool useDevelopment)
-               && useDevelopment)
-            {
-               return new AzureTableStorageKeyValueStorage();
-            }
-            else
-            {
-               connectionString.GetRequired(KnownParameter.AccountName, true, out string accountName);
-               connectionString.GetRequired(KnownParameter.KeyOrPassword, true, out string key);
-
-               return new AzureTableStorageKeyValueStorage(accountName, key);
-            }
-         }
 
          return null;
       }
