+++ src/storage.sln
@@ -57,6 +57,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blobs", "Azure\Storage.Net.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Files", "Azure\Storage.Net.Microsoft.Azure.Storage.Files\Files.csproj", "{598A07EE-E6D5-4B54-81FB-1C0639847B18}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tables", "Azure\Storage.Net.Microsoft.Azure.Storage.Tables\Tables.csproj", "{803CCAD9-3800-4354-B546-7D6123F8F3D3}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -127,6 +129,10 @@ Global
 		{598A07EE-E6D5-4B54-81FB-1C0639847B18}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{598A07EE-E6D5-4B54-81FB-1C0639847B18}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{598A07EE-E6D5-4B54-81FB-1C0639847B18}.Release|Any CPU.Build.0 = Release|Any CPU
+		{803CCAD9-3800-4354-B546-7D6123F8F3D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{803CCAD9-3800-4354-B546-7D6123F8F3D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{803CCAD9-3800-4354-B546-7D6123F8F3D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{803CCAD9-3800-4354-B546-7D6123F8F3D3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -152,6 +158,7 @@ Global
 		{8EED860F-4D6C-4EAE-97D3-62925E7336E0} = {E5D1DA80-E9FE-4965-A498-DFFC8ED70E40}
 		{6F4E9CCC-DEB7-4D40-B2BA-4914171D26B2} = {E5D1DA80-E9FE-4965-A498-DFFC8ED70E40}
 		{598A07EE-E6D5-4B54-81FB-1C0639847B18} = {E5D1DA80-E9FE-4965-A498-DFFC8ED70E40}
+		{803CCAD9-3800-4354-B546-7D6123F8F3D3} = {E5D1DA80-E9FE-4965-A498-DFFC8ED70E40}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C5FC6E9A-255E-4FD4-B08C-8EDE44BFE37A}
